Carzilla_Ecommerce_App
This is my very first project deployed using SpringBoot. It is an ecommerce app for cars, running on a local server and that is aimed at customers adding to carts, the admin having the ability to edit and delete a product at his own will. Admin also has the ability to register another admin.